# Optimized PDP calculation when multiple PDPs are closed and previous was open 