--- a/tests/Makefile.am
+++ b/tests/Makefile.am
@@ -4,7 +4,8 @@ TESTS = modify1 modify2 modify3 modify4 modify5 \
 	dump-restore \
 	create-with-source-1 create-with-source-2 create-with-source-3 \
 	create-with-source-4 create-with-source-and-mapping-1 \
-	create-from-template-1 dcounter1 vformatter1 xport1 list1
+	create-from-template-1 dcounter1 vformatter1 xport1 list1 \
+    pdp-calc1
 
 EXTRA_DIST = Makefile.am \
 	functions $(TESTS) \
@@ -17,7 +18,8 @@ EXTRA_DIST = Makefile.am \
 	tune1-testa-mod1.dump tune1-testa-mod2.dump tune1-testorg.dump \
 	tune2-testa-mod1.dump tune2-testorg.dump \
 	valgrind-supressions dcounter1 dcounter1.output graph1.output vformatter1 rpn1.output rpn2.output \
-	xport1.json.output xport1.xml.output
+	xport1.json.output xport1.xml.output \
+    pdp-calc1 pdp-calc1-1-avg-60.output pdp-calc1-1-avg-300.output pdp-calc1-1-max-300.output
 	
 # NB: AM_TESTS_ENVIRONMENT not available until automake 1.12
 AM_TESTS_ENVIRONMENT = \