If more than one PDP's are completed within an update and the previous chronologically oldest PDP was still open, there are two questionable effects:

1. The value of the last open PDP gets involved in the calculation of all PDPs to be completed. It is distributed evenly over them, even though it does not belong there in time:

**Example 1**
```
rrdtool create test.rrd --start 1548500640 -s 60 DS:AV:GAUGE:3600:0:1 RRA:AVERAGE:0.5:1:20
rrdtool update test.rrd 1548500700:1  # Initialize (value 1)
rrdtool update test.rrd 1548500730:1  # Force an open PDP (30s value 1)
rrdtool update test.rrd 1548500820:0  # Complete 2 PDP's  (90s value 0)

<!-- 2019-01-26 12:04:00 CET / 1548500640 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:05:00 CET / 1548500700 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:06:00 CET / 1548500760 --> <row><v>2.5000000000e-01</v></row>
<!-- 2019-01-26 12:07:00 CET / 1548500820 --> <row><v>2.5000000000e-01</v></row>
```
**Example 2**
```
rrdtool create test.rrd --start 1548500640 -s 60 DS:AV:GAUGE:3600:0:1 RRA:AVERAGE:0.5:1:20
rrdtool update test.rrd 1548500700:1  # Initialize (value 1)
rrdtool update test.rrd 1548500730:1  # Force an open PDP (30s value 1)
rrdtool update test.rrd 1548501000:0  # Complete 5 PDP's (270s value 0)

<!-- 2019-01-26 12:04:00 CET / 1548500640 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:05:00 CET / 1548500700 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:06:00 CET / 1548500760 --> <row><v>1.0000000000e-01</v></row>
<!-- 2019-01-26 12:07:00 CET / 1548500820 --> <row><v>1.0000000000e-01</v></row>
<!-- 2019-01-26 12:08:00 CET / 1548500880 --> <row><v>1.0000000000e-01</v></row>
<!-- 2019-01-26 12:09:00 CET / 1548500940 --> <row><v>1.0000000000e-01</v></row>
<!-- 2019-01-26 12:10:00 CET / 1548501000 --> <row><v>1.0000000000e-01</v></row>
```

2. If the previously open PDP is considered as invalid because it consists of more than half unknown step-seconds, all PDPs to be completed are classified as invalid. This explains the behavior in Issue #768:

**Example 1**
```
rrdtool create test.rrd --start 1548500640 -s 60 DS:AV:GAUGE:3600:0:1 RRA:AVERAGE:0.5:1:20
rrdtool update test.rrd 1548500700:1  # Initialize (value 1)
rrdtool update test.rrd 1548500731:U  # Force an open PDP with more than 50% unknown secs
rrdtool update test.rrd 1548500820:1  # Complete 2 PDP's (89s value 1)

<!-- 2019-01-26 12:05:00 CET / 1548500700 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:06:00 CET / 1548500760 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:07:00 CET / 1548500820 --> <row><v>NaN</v></row>
```
**Example 2**
```
rrdtool create test.rrd --start 1548500640 -s 60 DS:AV:GAUGE:3600:0:1 RRA:AVERAGE:0.5:1:20
rrdtool update test.rrd 1548500700:1  # Initialize (value 1)
rrdtool update test.rrd 1548500731:U  # Force an open PDP with more than 50% unknown secs
rrdtool update test.rrd 1548501000:1  # Complete 5 PDP's (270s value 1)

<!-- 2019-01-26 12:05:00 CET / 1548500700 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:06:00 CET / 1548500760 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:07:00 CET / 1548500820 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:08:00 CET / 1548500880 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:09:00 CET / 1548500940 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:10:00 CET / 1548501000 --> <row><v>NaN</v></row>
```
I am convinced that this is an error, because the already existing values of an open PDP do **not** affect the following PDP if this is not completed within the same update. So rrdtool works correctly if only 1 PDP is completed within one update:

**Example**
```
rrdtool create test.rrd --start 1548500640 -s 60 DS:AV:GAUGE:3600:0:1 RRA:AVERAGE:0.5:1:20
rrdtool update test.rrd 1548500700:1  # Initialize (value 1)
rrdtool update test.rrd 1548500730:1  # Force an open PDP (30s value 1)
rrdtool update test.rrd 1548500819:0  # Complete 1 PDP (89s value 0)
rrdtool update test.rrd 1548500820:0  # Complete 1 PDP (1s value 0)

<!-- 2019-01-26 12:05:00 CET / 1548500700 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:06:00 CET / 1548500760 --> <row><v>5.0000000000e-01</v></row>
<!-- 2019-01-26 12:07:00 CET / 1548500820 --> <row><v>0.0000000000e+00</v></row>
```
**Proposal for a simple solution**
If the situation described above occurs, the update can be split into two parts.
First, the previous PDP is calculated and completed, then all others. The update value is splitted proportionally to both parts. This solution solves both effects:

1. The existing values of the previous open PDP do not affect the calculation of the following PDPs any more:
```
rrdtool create test.rrd --start 1548500640 -s 60 DS:AV:GAUGE:3600:0:1 RRA:AVERAGE:0.5:1:20
rrdtool update test.rrd 1548500700:1  # Initialize (value 1)
rrdtool update test.rrd 1548500730:1  # Force open PDP (30s value 1)
rrdtool update test.rrd 1548501000:0  # Complete 5 PDP's (270s value 0)

<!-- 2019-01-26 12:04:00 CET / 1548500640 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:05:00 CET / 1548500700 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:06:00 CET / 1548500760 --> <row><v>5.0000000000e-01</v></row>
<!-- 2019-01-26 12:07:00 CET / 1548500820 --> <row><v>0.0000000000e+00</v></row>
<!-- 2019-01-26 12:08:00 CET / 1548500880 --> <row><v>0.0000000000e+00</v></row>
<!-- 2019-01-26 12:09:00 CET / 1548500940 --> <row><v>0.0000000000e+00</v></row>
<!-- 2019-01-26 12:10:00 CET / 1548501000 --> <row><v>0.0000000000e+00</v></row>
```
2. If the previously open PDP is considered invalid, the following PDP's are not affected anymore:
```
rrdtool create test.rrd --start 1548500640 -s 60 DS:AV:GAUGE:3600:0:1 RRA:AVERAGE:0.5:1:20
rrdtool update test.rrd 1548500700:1  # Initialize (value 1)
rrdtool update test.rrd 1548500731:U  # Force an open PDP with more than 50% unknown step-secs
rrdtool update test.rrd 1548501000:1  # Complete 5 PDP's

<!-- 2019-01-26 12:04:00 CET / 1548500640 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:05:00 CET / 1548500700 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:06:00 CET / 1548500760 --> <row><v>NaN</v></row>
<!-- 2019-01-26 12:07:00 CET / 1548500820 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:08:00 CET / 1548500880 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:09:00 CET / 1548500940 --> <row><v>1.0000000000e+00</v></row>
<!-- 2019-01-26 12:10:00 CET / 1548501000 --> <row><v>1.0000000000e+00</v></row>
```

What do you think?
Best regards
Jens