@oetiker 2019-02-12
Can you comment on how your new algorithm is to be viewed in light of a 'COUNTER' type datasource.
The curent design was primarily developed with COUNTER type datasources in minde.