@JKammler 2019-02-13
The COUNTER type seems to be fine. I will investigate and comment this later because I just found another issue with the algorithm. When write_to_rras() is called twice an RRA with multiple steps gets a bit confused in some situations... oh dear... hard stuff.