@oetiker 2019-02-13
make sure to add tests for the cases you are fixing