@oetiker 2019-03-03
any progress with testing ?