@JKammler 2019-03-04
Hi Tobi,

I already took a look at the other tests to get an idea how I have to do it and began with a setup. Unfortunately I did not find the time yet to finish it. But it is on the top of my Jira-List.

As you know I would be very happy if this issue could be fixed.



Von: Tobias Oetiker <notifications@github.com>
Gesendet: Sonntag, 3. März 2019 19:57
An: oetiker/rrdtool-1.x <rrdtool-1.x@noreply.github.com>
Cc: Kammler, Jens, (TV), NMD-A1 <Jens.Kammler@Bertelsmann.de>; Author <author@noreply.github.com>
Betreff: Re: [oetiker/rrdtool-1.x] Optimized PDP calculation when multiple PDPs are closed and previous was open (#970)



any progress with testing ?

—
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_oetiker_rrdtool-2D1.x_pull_970-23issuecomment-2D469052485&d=DwMCaQ&c=vo2ie5TPcLdcgWuLVH4y8lsbGPqIayH3XbK3gK82Oco&r=EcrkhjDuhLR6buT4aHGRlYqh6uQ4XfoYHMRXRqEmhEk&m=nxHS3KFEe6Z6DEIrInVsNaJNE7vvnyTKTsXfnDY039U&s=mpVuQWc9UR8ZgPlcy8VHmVwbutD9X9YFYTM8ZUj4NpM&e=>, or mute the thread<https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_notifications_unsubscribe-2Dauth_AYj8ZebREKSq5oCYyRBglhFGzo6HXoNXks5vTBr5gaJpZM4a3fAw&d=DwMCaQ&c=vo2ie5TPcLdcgWuLVH4y8lsbGPqIayH3XbK3gK82Oco&r=EcrkhjDuhLR6buT4aHGRlYqh6uQ4XfoYHMRXRqEmhEk&m=nxHS3KFEe6Z6DEIrInVsNaJNE7vvnyTKTsXfnDY039U&s=B-L_ng2Sw8QeHT0FpEGJpuDdsRl0mpTrMWEgxDeDcwI&e=>.