@JKammler 2019-03-06
Thank you for merging the code! Hopefully there is no negative side effect I didn’t consider.
BTW: While I was developing the tests, I noticed another minor issue independently of my changes with the unknown-seconds calculation for counter data sources… maybe I will report about and look into it when I find the time J